1019 hr2, seo_ju merge

+++ Project/src/main/webapp/Process/career/DeleteProcess.jsp
@@ -1,3 +1,4 @@
+<%@page import="like.likeBoardDAO"%>
 <%@page import="board2.careerBoardDTO"%>
 <%@page import="board2.careerBoardDAO"%>
 <%@page import="org.apache.tomcat.util.http.fileupload.FileUtils"%>
@@ -8,22 +9,24 @@
 String num = request.getParameter("pnum");
 careerBoardDAO dao = new careerBoardDAO();
 careerBoardDTO dto = dao.selectView(num);
-
+likeBoardDAO ldao = new likeBoardDAO();
 // 로그인된 사용자 ID 얻기
 String sessionId = session.getAttribute("UserId").toString();
 
 int delResult = 0;
+int likeResult = 0;
 
-// if(sessionId.equals(dto.getId())){
-	// 작성자 본인
 	dto.setNum(num);
-	System.out.println(num);
 	dao.posetdeleteCom(num);
 	dao.reportdelete(num);
-	delResult = dao.deletePost(dto);
-	dao.close();
 	
-	if(delResult == 1){
+	delResult = dao.deletePost(dto);
+	likeResult = ldao.deleteLike(Integer.parseInt(num), 2);
+	
+	dao.close();
+	ldao.close();
+	
+	if(delResult == 1 && likeResult > 0){
 		//성공 ) 목록 페이지로 이동
 		String FileName = dto.getFilename();
 		dao.deleteFile(request, "/Uploads", FileName);
